+++ src/kernel/mmio.c
@@ -6,20 +6,20 @@
 #include "../../include/common/stdint.h"
 
 /* Memory-Mapped I/O output */
-void mmio_write(uint32_t reg, uint32_t data)
+void mmio_write(size_t reg, size_t data)
 {
-    *(volatile uint32_t *)reg = data;
+    *(volatile size_t *)reg = data;
 }
 
-void mmio_write_with_mask(uint32_t reg, volatile void *buffer, uint32_t mask)
+void mmio_write_with_mask(size_t reg, volatile void *buffer, uint32_t mask)
 {
-    mmio_write(reg, *((uint32_t *)buffer) & mask);
+    mmio_write(reg, *((size_t *)buffer) & mask);
 }
 
-void mmio_write_out(uint32_t reg, volatile void *buffer, size_t words)
+void mmio_write_out(size_t reg, volatile void *buffer, size_t words)
 {
-    volatile uint32_t *word_aligned_buffer = (uint32_t *)buffer;
-    volatile uint32_t *memory_address = (uint32_t *)reg;
+    volatile size_t *word_aligned_buffer = (size_t *)buffer;
+    volatile size_t *memory_address = (size_t *)reg;
     for (; words; words--)
     {
         *memory_address++ = *word_aligned_buffer++;
@@ -27,15 +27,15 @@ void mmio_write_out(uint32_t reg, volatile void *buffer, size_t words)
 }
 
 /* Memory-Mapped I/O input */
-uint32_t mmio_read(uint32_t reg)
+uint32_t mmio_read(size_t reg)
 {
-    return *(volatile uint32_t *)reg;
+    return *(volatile size_t *)reg;
 }
 
-void mmio_read_in(uint32_t reg, volatile void *buffer, size_t words)
+void mmio_read_in(size_t reg, volatile void *buffer, size_t words)
 {
-    volatile uint32_t *word_aligned_buffer = (uint32_t *)buffer;
-    volatile uint32_t *memory_address = (uint32_t *)reg;
+    volatile size_t *word_aligned_buffer = (size_t *)buffer;
+    volatile size_t *memory_address = (size_t *)reg;
     for (; words; words--)
     {
         *word_aligned_buffer++ = *memory_address++;
